requirements.txt
pydub
ruamel.yaml
requests
#genanki
-e git://github.com/mi-ael/genanki#egg=genanki
cachier
num2words
googledrivedownloader